requirements.txt
aiogrpc==1.8
altgraph @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.4.0
autogen==1.0.16
autopep8==2.3.1
backports.tarfile==1.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
distro==1.9.0
docopt==0.6.2
docutils==0.21.2
exceptiongroup==1.2.1
filelock==3.15.4
fsspec==2024.6.1
future @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.0.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
Jinja2==3.1.4
keyring==25.2.1
llvmlite==0.43.0
macholib @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mavsdk==2.8.0
mdurl==0.1.2
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.2.1
nh3==0.2.17
numba==0.60.0
numpy==2.0.0
openai==1.35.9
openai-whisper==20231117
pkginfo==1.10.0
protobuf==3.20.1
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.1
readme_renderer==43.0
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
scipy==1.13.1
six @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
sounddevice==0.4.7
sympy==1.12.1
tiktoken==0.7.0
tomli==2.0.1
torch==2.3.1
tqdm==4.66.4
twine==5.1.1
typing_extensions==4.12.2
urllib3==2.2.2
zipp==3.19.2
